requirements.txt
functions-framework==3.*
google-cloud-kms
google-auth
google-api-python-client